server/requirements.txt
openenv-core[core]>=0.2.2
uvicorn>=0.29.0
fastapi>=0.110.0
